Some rearrangement and documentation
