+++ HL_kent_binaries/readme.txt
@@ -1,6 +1,3 @@
-chainAntiRepeat
-chainCleaner
-chainMergeSort
-chainScore
-chainSort
-pslSortAcc
+Directory to store Kent binaries necessary to run the pipeline,
+which are not included in the $PATH for some reason,
+and were downloaded using install_dependencies.py script.
