--- standalone_scripts/readme.txt
@@ -1,3 +1,0 @@
-All perl scripts are to be translated into Python and refactored:
-
--
